+++ src/main/resources/test-applicationContext.xml
@@ -5,6 +5,10 @@
 
     <bean id="userDao" class="springbook.user.dao.UserDao">
         <property name="dataSource" ref="dataSource"/>
+        <property name="jdbcContext" ref="jdbcContext"/>
+    </bean>
+    <bean id="jdbcContext" class="springbook.user.dao.JdbcContext">
+        <property name="dataSource" ref="dataSource"/>
     </bean>
     <bean id="dataSource" class="org.springframework.jdbc.datasource.SimpleDriverDataSource">
         <property name="driverClass" value="com.mysql.cj.jdbc.Driver"/>
